examples/maintenance_timer/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform

azure-functions
azure-functions-langgraph[azure-table]>=0.6.0,<0.7.0
